requirements.txt
flask
openai
azure-storage-blob
python-dotenv
